Move agreement states to validators
