+++ packages/agreement-process/src/model/domain/models.ts
@@ -2,7 +2,6 @@ import {
   AgreementAttribute,
   AgreementStamps,
   AgreementState,
-  agreementState,
 } from "pagopa-interop-models";
 import { z } from "zod";
 
@@ -49,63 +48,3 @@ export const CompactEService = z.object({
   name: z.string(),
 });
 export type CompactEService = z.infer<typeof CompactEService>;
-
-export const agreementActivableStates: AgreementState[] = [
-  agreementState.pending,
-  agreementState.suspended,
-];
-export const agreementSuspendableStates: AgreementState[] = [
-  agreementState.active,
-  agreementState.suspended,
-];
-export const agreementArchivableStates: AgreementState[] = [
-  agreementState.active,
-  agreementState.suspended,
-];
-export const agreementSubmittableStates: AgreementState[] = [
-  agreementState.draft,
-];
-
-export const agreementUpdatableStates: AgreementState[] = [
-  agreementState.draft,
-];
-
-export const agreementUpgradableStates: AgreementState[] = [
-  agreementState.active,
-  agreementState.suspended,
-];
-export const agreementRejectableStates: AgreementState[] = [
-  agreementState.pending,
-];
-
-export const agreementDeletableStates: AgreementState[] = [
-  agreementState.draft,
-  agreementState.missingCertifiedAttributes,
-];
-
-export const agreementActivationFailureStates: AgreementState[] = [
-  agreementState.draft,
-  agreementState.pending,
-  agreementState.missingCertifiedAttributes,
-];
-
-export const agreementCloningConflictingStates: AgreementState[] = [
-  agreementState.draft,
-  agreementState.pending,
-  agreementState.missingCertifiedAttributes,
-  agreementState.active,
-  agreementState.suspended,
-];
-
-export const agreementCreationConflictingStates: AgreementState[] = [
-  agreementState.draft,
-  agreementState.pending,
-  agreementState.missingCertifiedAttributes,
-  agreementState.active,
-  agreementState.suspended,
-];
-
-export const agreementSubmissionConflictingStates: AgreementState[] = [
-  agreementState.pending,
-  agreementState.missingCertifiedAttributes,
-];
